Readme edits
I added some additional info for running with Yarn. Although it is already there ( not touched ) I took it our of the Development section since it may look intimidating for non-devs. 

I also made a modification in the example with the intention of showing the simple hello world and another example that uses 2 events.